Added BluetoothDiscovery class
Had moved logic from main.cpp to 2 separate methods named startDeviceDiscovery and deviceDiscovered for logging. Had faced issues to use signal and slots mechanism, majorly due to the Q_OBJECT macro not added, so it wasn't able to detect it.

The above class does detect nearby bluetooth devices. However, Sony headphone device showing in LE-WF1000XM3 so completely not working. Need to dig in but have no clue...

Hoping to find a solution to this...